Refactored nvi-code into new file structure (#8346)
* Moved two pages and two hooks

* Moved third page

* Moved fourth page

* Moved cell styling

* Fixed import

* Moved sort helpers into common sort helper file

* Moved admin sort type to separate file

* New hook for both filtering and sorting

* New folder naming

* '-' between words in all files but hooks and components
