+++ src/components/sort-selectors/sort-nvi-table/nvi-admin-sort-helpers.ts
@@ -1,6 +1,29 @@
-import { SortSelectorOption } from '../../../../../components/_molecules/SortSelector';
-import { NviAdminOrderBy } from '../nviAdminHelpers';
-import { NviAdminSortSelectorType } from './NviAdminSortSelector';
+import { TFunction } from 'i18next';
+import {
+  getNviApprovedByEverybody,
+  getNviApprovedCount,
+  getNviCandidatesCount,
+  getNviInstitutionName,
+  getNviRejectedCount,
+  getNviSectorLabel,
+  getNviTotalCount,
+  getNviValidPoints,
+} from '../../../pages/basic_data/app_admin/nviAdmin/nviAdminHelpers';
+import { InstitutionReport } from '../../../types/nvi.types';
+import { SortSelectorOption } from '../../_molecules/SortSelector';
+import { NviAdminSortSelectorType } from './nvi-admin-sort-types';
+
+export enum NviAdminOrderBy {
+  Institution = 'institution',
+  Sector = 'sector',
+  Points = 'points',
+  CandidatesApprovedByInstitution = 'candidatesApprovedByInstitution',
+  CandidatesApprovedByEverybody = 'candidatesApprovedByEverybody',
+  TotalNumber = 'totalNumber',
+  NumberOfCandidates = 'numberOfCandidates',
+  NumberOfApprovedCandidates = 'numberOfApprovedCandidates',
+  NumberOfRejectedCandidates = 'numberOfRejectedCandidates',
+}
 
 const commonOptions: SortSelectorOption[] = [
   {
@@ -109,5 +132,29 @@ export const getNviAdminSortOptions = (type: NviAdminSortSelectorType) => {
       return [...commonOptions, ...pointsOptions];
     default:
       return [...commonOptions];
+  }
+};
+export const getNviAdminSortValue = (report: InstitutionReport, orderBy: NviAdminOrderBy, t: TFunction) => {
+  switch (orderBy) {
+    case NviAdminOrderBy.Institution:
+      return getNviInstitutionName(report).toLowerCase();
+    case NviAdminOrderBy.Sector:
+      return getNviSectorLabel(report, t).toLowerCase();
+    case NviAdminOrderBy.Points:
+      return getNviValidPoints(report);
+    case NviAdminOrderBy.CandidatesApprovedByInstitution:
+      return getNviApprovedCount(report);
+    case NviAdminOrderBy.CandidatesApprovedByEverybody:
+      return getNviApprovedByEverybody(report);
+    case NviAdminOrderBy.TotalNumber:
+      return getNviTotalCount(report);
+    case NviAdminOrderBy.NumberOfCandidates:
+      return getNviCandidatesCount(report);
+    case NviAdminOrderBy.NumberOfApprovedCandidates:
+      return getNviApprovedCount(report);
+    case NviAdminOrderBy.NumberOfRejectedCandidates:
+      return getNviRejectedCount(report);
+    default:
+      return getNviInstitutionName(report).toLowerCase();
   }
 };
