+++ src/components/sort-selectors/sort-nvi-table/nvi-admin-sort-types.ts
@@ -1,0 +1,4 @@
+export enum NviAdminSortSelectorType {
+  Points = 'points',
+  Status = 'status',
+}
