+++ src/hooks/nvi/useSortInstitutionReportsByUrl.ts
@@ -1,9 +1,12 @@
 import { useTranslation } from 'react-i18next';
 import { useSearchParams } from 'react-router';
-import { getNviAdminSortValue, NviAdminOrderBy } from '../../pages/basic_data/app_admin/nviAdmin/nviAdminHelpers';
+import {
+  getNviAdminSortValue,
+  NviAdminOrderBy,
+} from '../../components/sort-selectors/sort-nvi-table/nvi-admin-sort-helpers';
 import { InstitutionReport } from '../../types/nvi.types';
 
-export const useSortInstitutionReports = (reports: InstitutionReport[]) => {
+export const useSortInstitutionReportsByUrl = (reports: InstitutionReport[]) => {
   const [searchParams] = useSearchParams();
   const { t } = useTranslation();
