+++ src/pages/basic_data/app_admin/nviAdmin/NviAdminPublicationPointsRow.tsx
@@ -1,7 +1,8 @@
 import { TableCell, TableRow } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 import { PercentageWithIcon } from '../../../../components/_atoms/PercentageWithIcon';
-import { CenteredTableCell, HorizontalBox } from '../../../../components/styled/Wrappers';
+import { HorizontalBox } from '../../../../components/styled/Wrappers';
+import { CenteredTableCell } from '../../../../styles/table-styles';
 import { InstitutionReport } from '../../../../types/nvi.types';
 import {
   getNviApprovedByEverybody,
