+++ src/pages/messages/TasksPage.tsx
@@ -27,12 +27,12 @@ import { NviCandidatesList } from './components/NviCandidatesList';
 import { NviCandidatesNavigationAccordion } from './components/NviCandidatesNavigationAccordion';
 import { NviCorrectionList } from './components/NviCorrectionList';
 import { NviCorrectionListNavigationAccordion } from './components/NviCorrectionListNavigationAccordion';
-import { NviStatusPage } from './components/NviStatusPage';
+import { NviStatusPage } from '../tasks/nvi/status/NviStatusPage';
 import { ResultRegistrationsNavigationListAccordion } from './components/ResultRegistrationsNavigationListAccordion';
 import { TicketList } from './components/TicketList';
 import { TicketTypeTag } from './components/TicketTypeTag';
 import { NviDisputePage } from './components/NviDisputePage';
-import { NviPublicationPointsPage } from './components/NviPublicationPointsPage';
+import { NviPublicationPointsPage } from '../tasks/nvi/publication-points/NviPublicationPointsPage';
 import { useFetchTickets } from '../../api/hooks/useFetchTickets';
 import { useGetNotificationCounts, useGetTicketsCounts } from './user-dialog-helpers';
 import { checkPages } from './tasks-helpers';
