+++ src/styles/table-styles.ts
@@ -1,0 +1,5 @@
+import { styled, TableCell } from '@mui/material';
+
+export const CenteredTableCell = styled(TableCell)({
+  textAlign: 'center',
+});
